requirements.txt
cmake
dlib
face-recognition
cvzone
opencv-python
numpy
firebase-admin
python-dotenv